feature: Add likes count & get liked list (goods, store)

+++ src/main/java/com/jocoos/mybeautip/store/StoreRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.Query;
 public interface StoreRepository extends JpaRepository<Store, Long> {
   @Modifying
   @Query("update Store s set s.likeCount = s.likeCount + ?2, s.modifiedAt = now() where s.id = ?1")
-  void updateLikeCount(Long id, Long count);
+  void updateLikeCount(Long id, Integer count);
 }
